Make 'additional_security_group_ids' optional
As described in the issue in the reference down below, the creation of a
custom worker pool fails when there is no 'additional_security_group_ids'
specified, when it points to an empty list or when the list contains
only an empty string (as in the examples in this repository).

The Red Hat's terraform-provider-rhcs provider [1] defines this attribute
as optional:

    "additional_security_group_ids": {
        "type": [
          "list",
          "string"
        ],
        "description": "Additional security group ids. After the creation of the resource, it is not possible to update the attribute value.",
        "description_kind": "plain",
        "optional": true
      }

This commit changes the 'aws_node_pool' variable to reflect the optional
nature of this attribute; after applying this change, the custom pool
creation is successful.

[1] https://github.com/terraform-redhat/terraform-provider-rhcs

Fixes #78