WIP: Support ROSA GovCloud HCP clusters
Before this commit, the modules made two assumptions which prevented creation
of HCP clusters on ROSA GovCloud:

1. ARN formats hard-coded the AWS partition
2. The `rhcs_cluster_rosa_hcp` module always set a billing account ID, which should
   always be empty for GovCloud HCP clusters.

This commit makes ARN partition segments dynamic with regards to the current AWS
account, and sets a `null` billing account ID when targeting GovCloud. The result
is HCP clusters can be successfully created on ROSA GovCloud.

TODO: This depends on a new provider version which includes:

- https://github.com/terraform-redhat/terraform-provider-rhcs/pull/946
- https://github.com/terraform-redhat/terraform-provider-rhcs/pull/947